api-rest/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
python-dotenv>=0.21.0
firebase-admin>=6.0.0
google-cloud-storage>=2.0.0
gunicorn>=20.1.0
requests>=2.28.0
pymysql>=1.0.2